+++ Modules/Intent.Modules.Application.ServiceImplementations.CRUD/MethodImplementationStrategies/UpdateImplementationStrategy.cs
@@ -83,8 +83,14 @@ namespace Intent.Modules.Application.ServiceImplementations.Conventions.CRUD.Met
                 .AddStatement($@"throw new {_template.GetNotFoundExceptionName()}($""Could not find {domainModel.Name.ToPascalCase()} {{{idParam}}}"");"));
             codeLines.AddRange(GetDTOPropertyAssignments($"existing{domainTypePascalCased}", dtoParam.Name, domainModel.InternalElement, dtoModel.Fields, true));
 
+            if (RepositoryRequiresExplicitUpdate(domainModel))
+            {
+                codeLines.Add($"{repositoryFieldName.ToPrivateMemberName()}.Update(existing{domainTypePascalCased});");
+            }
+
             if (operationModel.TypeReference.Element.IsDTOModel())
             {
+                codeLines.Add($"await {repositoryFieldName.ToPrivateMemberName()}.UnitOfWork.SaveChangesAsync(cancellationToken);");
                 codeLines.Add($"return existing{domainTypePascalCased}.MapTo{_template.GetTypeName((IElement)operationModel.TypeReference.Element)}(_mapper);");
             }
 
@@ -110,6 +116,17 @@ namespace Intent.Modules.Application.ServiceImplementations.Conventions.CRUD.Met
                 ctor.AddParameter(_template.UseType("AutoMapper.IMapper"), "mapper", parm => parm.IntroduceReadonlyField());
             }
         }
+
+        private bool RepositoryRequiresExplicitUpdate(ClassModel domainModel)
+        {
+            return _template.TryGetTemplate<ICSharpFileBuilderTemplate>(
+                       TemplateFulfillingRoles.Repository.Interface.Entity,
+                       domainModel,
+                       out var repositoryInterfaceTemplate) &&
+                   repositoryInterfaceTemplate.CSharpFile.Interfaces[0].TryGetMetadata<bool>("requires-explicit-update", out var requiresUpdate) &&
+                   requiresUpdate;
+        }
+
 
         private IList<CSharpStatement> GetDTOPropertyAssignments(string entityVarName, string dtoVarName, IElement domainModel, IList<DTOFieldModel> dtoFields, bool skipIdField)
         {