+++ Tests/GraphQL.MongoDb.TestApplication/GraphQL.MongoDb.TestApplication.Application/Privileges/UpdatePrivilege/UpdatePrivilegeCommandHandler.cs
@@ -37,6 +37,9 @@ namespace GraphQL.MongoDb.TestApplication.Application.Privileges.UpdatePrivilege
             }
             existingPrivilege.Name = request.Name;
             existingPrivilege.Description = request.Description;
+
+            _privilegeRepository.Update(existingPrivilege);
+            await _privilegeRepository.UnitOfWork.SaveChangesAsync(cancellationToken);
             return existingPrivilege.MapToPrivilegeDto(_mapper);
         }
     }