+++ Tests/GraphQL.MongoDb.TestApplication/GraphQL.MongoDb.TestApplication.Application/Users/UpdateUser/UpdateUserCommandHandler.cs
@@ -38,6 +38,9 @@ namespace GraphQL.MongoDb.TestApplication.Application.Users.UpdateUser
             existingUser.Name = request.Name;
             existingUser.Surname = request.Surname;
             existingUser.Email = request.Email;
+
+            _userRepository.Update(existingUser);
+            await _userRepository.UnitOfWork.SaveChangesAsync(cancellationToken);
             return existingUser.MapToUserDto(_mapper);
         }
     }